ci(test): change to vitest
