+++ package.json
@@ -14,7 +14,8 @@
 	},
 	"devDependencies": {
 		"@biomejs/biome": "1.9.1",
-		"turbo": "^2.1.2"
+		"turbo": "^2.1.2",
+		"vitest": "^2.1.2"
 	},
 	"packageManager": "pnpm@9.7.0",
 	"engines": {
