+++ packages/cooldown/package.json
@@ -18,7 +18,7 @@
 		"lint": "biome lint --write ./src",
 		"format": "biome format --write ./src",
 		"checkb": "biome check --write --no-errors-on-unmatched ./src",
-		"test": "node --test ./test/*"
+		"test": "vitest run --config ./test/vitest.config.ts ./test/"
 	},
 	"devDependencies": {
 		"@types/node": "^22.5.5",
