+++ packages/cooldown/test/manager.test.ts
@@ -1,0 +1,103 @@
+import { Cache, Client, Logger, MemoryAdapter } from 'seyfert';
+import { CommandHandler } from 'seyfert/lib/commands/handler.js';
+import { assert, beforeEach, describe, test } from 'vitest';
+import { CooldownManager, type CooldownProps, CooldownType } from '../src';
+
+describe('CooldownManager', async () => {
+	let client: Client;
+	let cooldownManager: CooldownManager;
+	let cooldownData: CooldownProps;
+
+	beforeEach(() => {
+		client = new Client({
+			getRC: () => ({
+				debug: true,
+				intents: 0,
+				token: '',
+				locations: {
+					base: '',
+					output: '',
+				},
+			}),
+		});
+
+		const handler = new CommandHandler(new Logger({ active: true }), client);
+		cooldownData = {
+			type: CooldownType.User,
+			interval: 1000,
+			uses: 3,
+		};
+		handler.values = [
+			// @ts-expect-error
+			{
+				name: 'testCommand',
+				cooldown: cooldownData,
+			},
+		];
+
+		client.commands = handler;
+
+		client.cache = new Cache(0, new MemoryAdapter(), {}, client);
+		cooldownManager = new CooldownManager(client);
+	});
+
+	await test('Data should return cooldown data for a command', () => {
+		const data = cooldownManager.getCommandData('testCommand');
+		assert.deepEqual(data, {
+			type: CooldownType.User,
+			interval: 1000,
+			uses: 3,
+		});
+	});
+
+	await test('Data should return undefined for non-existent command', () => {
+		const data = cooldownManager.getCommandData('nonExistentCommand');
+		assert.equal(data, undefined);
+	});
+
+	await test('has should return false for a new cooldown', () => {
+		const result = cooldownManager.has('testCommand', 'user1');
+		assert.equal(result, false);
+	});
+
+	await test('has should return true when cooldown is active', () => {
+		for (let i = 0; i < cooldownData.uses; i++) {
+			cooldownManager.use('testCommand', 'user1');
+		}
+		const result = cooldownManager.has('testCommand', 'user1');
+		assert.equal(result, true);
+	});
+
+	await test('use should set cooldown when used for the first time', () => {
+		const result = cooldownManager.use('testCommand', 'user2');
+		assert.equal(result, true);
+	});
+
+	await test('use should return time left when cooldown is active', () => {
+		for (let i = 0; i < cooldownData.uses; i++) {
+			cooldownManager.use('testCommand', 'user3');
+		}
+		const result = cooldownManager.use('testCommand', 'user3');
+		assert.ok(typeof result === 'number');
+	});
+
+	await test('refill should refill the cooldown', () => {
+		cooldownManager.use('testCommand', 'user1');
+		const result = cooldownManager.refill('testCommand', 'user1');
+		assert.equal(result, true);
+		assert.equal(cooldownManager.has('testCommand', 'user1'), false);
+	});
+
+	await test('drip should drip the cooldown over time', async () => {
+		cooldownManager.use('testCommand', 'user1');
+
+		// Simulate time passing
+		await new Promise(resolve => setTimeout(resolve, 1000));
+
+		const data = cooldownManager.resource.get('testCommand:user:user1');
+		const props = cooldownManager.getCommandData('testCommand');
+		await cooldownManager.drip('testCommand', 'user1', props!, data!);
+		const getter = cooldownManager.resource.get('testCommand:user:user1');
+		assert.ok(getter?.remaining === 2);
+	});
+});
