+++ packages/redis-adapter/package.json
@@ -17,7 +17,7 @@
 		"build": "tsc",
 		"lint": "biome lint --write ./src",
 		"format": "biome format --write ./src",
-		"test": "node --test ./test/*",
+		"test": "vitest run --config ./test/vitest.config.ts ./test/",
 		"checkb": "biome check --write --no-errors-on-unmatched ./src",
 		"test:adapter": "node --test ./test/adapter.test.js",
 		"test:cache": "node --test ./test/cache.test.js",
