--- packages/redis-adapter/test/cache.test.js
@@ -1,35 +1,0 @@
-const { describe, test } = require('node:test');
-const { Client } = require('seyfert');
-const { RedisAdapter } = require('../lib/index');
-
-// all intents
-const intents = 53608447;
-
-describe('Test Adapter cache', async t => {
-	const adapter = new RedisAdapter({
-		redisOptions: {},
-		namespace: 'test_cache',
-	});
-
-	await adapter.start();
-
-	test('discord cache', async () => {
-		const client = new Client({
-			getRC: async () => ({
-				locations: {
-					base: '',
-					output: '',
-				},
-				intents,
-				token: '',
-			}),
-		});
-		client.setServices({
-			cache: {
-				adapter,
-			},
-		});
-		await client.cache.testAdapter();
-		await adapter.client.quit();
-	});
-});
