+++ packages/redis-adapter/test/vitest.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+	test: {
+		fileParallelism: false,
+		isolate: false,
+	},
+});
